moved post components into posts directory
